fix: small fixes and refactoring
- Rename util functions
- Update flush function, move it to the tests folder
- Rename types
- Refactor controllers execution on the routes
- Refactor repository query result types
