+++ src/routers/campaign.router.ts
@@ -1,5 +1,4 @@
 import express from 'express';
-import type { Request, Response, NextFunction } from 'express';
 import campaignController from '../controllers/campaign.controller';
 import validate from '../middlewares/schema-validator';
 import { CreateCampaignSchema, ListCampaignSchema } from '../controllers/campaign-controller.schema';
@@ -7,11 +6,9 @@ import { CreateCampaignSchema, ListCampaignSchema } from '../controllers/campaig
 export default function campaignRouter() {
   const router = express.Router();
 
-  router.post('/', validate(CreateCampaignSchema), async (req: Request, res: Response, next: NextFunction) =>
-    campaignController.create(req, res, next),
-  );
+  router.post('/', validate(CreateCampaignSchema), campaignController.create);
 
-  router.get('/', validate(ListCampaignSchema), async (req: Request, res: Response, next: NextFunction) => campaignController.list(req, res, next));
+  router.get('/', validate(ListCampaignSchema), campaignController.list);
 
   return router;
 }
